requirements.txt
numpy~=1.26.4
pillow~=10.2.0
scipy~=1.12.0
matplotlib~=3.8.2
pydantic~=2.6.1
pymongo~=4.6.1
requests~=2.31.0
python-dotenv~=1.0.1